research/requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
Faker==30.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
